Time: 24 ms (40.61%), Space: 6 MB (31.89%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-    long long int temp, revers=0, rem;
+    long int temp, revers=0, rem;
     temp = x;
     if(x < 0 || (x != 0 && x % 10 == 0))
         return false;
